Remove the size limit for memory read and write (revamped) (#2144)
* modify uc_priv.h to use 64 bit hwaddr type for length for uc_write_mem_t and uc_read_mem_t

* propogate 64 bit address support to memory function inside unicorn.h

* modify physical memory functions to ensure 64 bit support for size parameters

* apply  uint64_t size handling inside the memory handling main functions

* remove INT_MAX size check as we now use a 64 bit address

* add @secretnonempty unit test for large memory addresses to avoid regression

* replace size_t in python3 bindings with u64

* adjust rust bindings

* use format.sh script to improve formatting

* small adjustments to java bindings, because jlong type is sometimes already used

* adjust dotnet bindings to reflect new type sizes

* adjust pascal bindings

* adjust ruby bindings

* adjust vb6 bindings

* adjust zig bindings for 64 bit size support

* fix build.zig.zon for newer zig versions

* add haskell binding support for 64 bit sizes

* use mlugg/setup-zig instead of goto-bus-stop/setup-zig github action to get access to current zig releases

* use uint64_t instead of size_t for uc_mem_unmap

* add whitespace between functions to preserver formatting

* add precompiler check for android and abort execution of testcase to avoid infinite low memory killer loop

* downgrade zig version to 0.14.0 again to ensure that pipeline can run through

* adjust rust mmio_unwrap type to use u64

* change ffi size type of uc_mem_map_ptr into u64

* add warning to rust slicing

* return read_mem_as_vec to earlier state, leave usize as argument because there cannot be larger vec![] anyway

* adjust regions inside MmioCallbackScope to use u64 for region sizes

* adjust regions inside MmioCallbackScope to use u64 for region sizes inside search vec

* finalize dotnet bindings inside IBinding.fs

* fix typo

* adjust types inside Unicorn.fs

* finalize type wrangling inside F# bindings, setting MemRead and MemWrite to correct types

* fix missing bracket inside bindings/rust/src/lib.rs

* replace ffi interface to hook interface inside rust lib

* fix testcases to use u64 for longer references

+++ bindings/go/unicorn/unicorn.go
@@ -187,14 +187,14 @@ func (u *uc) MemWrite(addr uint64, data []byte) error {
 	if len(data) == 0 {
 		return nil
 	}
-	return errReturn(C.uc_mem_write(u.handle, C.uint64_t(addr), unsafe.Pointer(&data[0]), C.size_t(len(data))))
+	return errReturn(C.uc_mem_write(u.handle, C.uint64_t(addr), unsafe.Pointer(&data[0]), C.uint64_t(len(data))))
 }
 
 func (u *uc) MemReadInto(dst []byte, addr uint64) error {
 	if len(dst) == 0 {
 		return nil
 	}
-	return errReturn(C.uc_mem_read(u.handle, C.uint64_t(addr), unsafe.Pointer(&dst[0]), C.size_t(len(dst))))
+	return errReturn(C.uc_mem_read(u.handle, C.uint64_t(addr), unsafe.Pointer(&dst[0]), C.uint64_t(len(dst))))
 }
 
 func (u *uc) MemRead(addr, size uint64) ([]byte, error) {
@@ -203,7 +203,7 @@ func (u *uc) MemRead(addr, size uint64) ([]byte, error) {
 }
 
 func (u *uc) MemMapProt(addr, size uint64, prot int) error {
-	return errReturn(C.uc_mem_map(u.handle, C.uint64_t(addr), C.size_t(size), C.uint32_t(prot)))
+	return errReturn(C.uc_mem_map(u.handle, C.uint64_t(addr), C.uint64_t(size), C.uint32_t(prot)))
 }
 
 func (u *uc) MemMap(addr, size uint64) error {
@@ -211,15 +211,15 @@ func (u *uc) MemMap(addr, size uint64) error {
 }
 
 func (u *uc) MemMapPtr(addr, size uint64, prot int, ptr unsafe.Pointer) error {
-	return errReturn(C.uc_mem_map_ptr(u.handle, C.uint64_t(addr), C.size_t(size), C.uint32_t(prot), ptr))
+	return errReturn(C.uc_mem_map_ptr(u.handle, C.uint64_t(addr), C.uint64_t(size), C.uint32_t(prot), ptr))
 }
 
 func (u *uc) MemProtect(addr, size uint64, prot int) error {
-	return errReturn(C.uc_mem_protect(u.handle, C.uint64_t(addr), C.size_t(size), C.uint32_t(prot)))
+	return errReturn(C.uc_mem_protect(u.handle, C.uint64_t(addr), C.uint64_t(size), C.uint32_t(prot)))
 }
 
 func (u *uc) MemUnmap(addr, size uint64) error {
-	return errReturn(C.uc_mem_unmap(u.handle, C.uint64_t(addr), C.size_t(size)))
+	return errReturn(C.uc_mem_unmap(u.handle, C.uint64_t(addr), C.uint64_t(size)))
 }
 
 func (u *uc) Query(queryType int) (uint64, error) {
